requirements.txt
alembic==1.13.2
amqp==5.2.0
billiard==4.2.0
blinker==1.8.2
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.1
Mako==1.3.5
MarkupSafe==2.1.5
prompt_toolkit==3.0.47
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.8
setuptools==75.1.0
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.4
wheel==0.44.0
